first steps in migrating to react-router v6
