+++ grafana-plugin/src/pages/incident/Incident.tsx
@@ -25,7 +25,6 @@ import { observer } from 'mobx-react';
 import moment from 'moment-timezone';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import Emoji from 'react-emoji-render';
-import { RouteComponentProps, withRouter } from 'react-router-dom';
 import reactStringReplace from 'react-string-replace';
 import { Colors, getLabelBackgroundTextColorObject } from 'styles/utils.styles';
 import { OnCallPluginExtensionPoints } from 'types';
@@ -69,10 +68,16 @@ import { parseURL } from 'utils/url';
 import { openNotification } from 'utils/utils';
 
 import { getActionButtons } from './Incident.helpers';
+import { PropsWithRouter, withRouter } from 'utils/hoc';
+import { getIncidentStyles } from './incident.styles';
 
 const INTEGRATION_NAME_LENGTH_LIMIT = 30;
 
-interface IncidentPageProps extends WithStoreProps, PageProps, RouteComponentProps<{ id: string }> {
+interface RouteProps {
+  id: string;
+}
+
+interface IncidentPageProps extends WithStoreProps, PageProps, PropsWithRouter<RouteProps> {
   theme: GrafanaTheme2;
 }
 
@@ -102,7 +107,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
   }
 
   componentDidUpdate(prevProps: IncidentPageProps) {
-    if (this.props.match.params.id !== prevProps.match.params.id) {
+    if (this.props.router.params.id !== prevProps.router.params.id) {
       this.update();
     }
   }
@@ -112,7 +117,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
 
     const {
       store,
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
@@ -128,7 +133,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
     const {
       store,
       query: { cursor, start, perpage },
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
@@ -136,7 +141,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
     const { errorData, showIntegrationSettings, showAttachIncidentForm, silenceModalData } = this.state;
     const { isNotFoundError, isWrongTeamError, isUnknownError } = errorData;
     const { alerts } = store.alertGroupStore;
-    const styles = getStyles(this.props.theme);
+    const styles = getIncidentStyles(this.props.theme);
 
     const incident = alerts.get(id);
 
@@ -274,7 +279,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
   handlePagedUserRemove = (userId: ApiSchemas['User']['pk']) => {
     return async () => {
       const {
-        match: {
+        router: {
           params: { id: alertId },
         },
       } = this.props;
@@ -289,12 +294,13 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
     const {
       store,
       query,
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
+
     const { alerts } = store.alertGroupStore;
-    const styles = getStyles(this.props.theme);
+    const styles = getIncidentStyles(this.props.theme);
 
     const incident = alerts.get(id);
     const integration = AlertReceiveChannelHelper.getIntegrationSelectOption(
@@ -518,13 +524,13 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
   renderTimeline = () => {
     const {
       store,
-      match: {
+      router: {
         params: { id },
       },
       theme,
     } = this.props;
 
-    const styles = getStyles(theme);
+    const styles = getIncidentStyles(theme);
     const incident = store.alertGroupStore.alerts.get(id);
 
     if (!incident.render_after_resolve_report_json) {
@@ -634,7 +640,7 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
   handleCreateResolutionNote = async () => {
     const {
       store,
-      match: {
+      router: {
         params: { id },
       },
     } = this.props;
@@ -708,7 +714,8 @@ class _IncidentPage extends React.Component<IncidentPageProps, IncidentPageState
 }
 
 function Incident({ incident }: { incident: ApiSchemas['AlertGroup']; datetimeReference: string }) {
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getIncidentStyles);
+
   return (
     <div key={incident.pk}>
       <div
@@ -732,7 +739,7 @@ function GroupedIncidentsList({
 }) {
   const store = useStore();
   const incident = store.alertGroupStore.alerts.get(id);
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getIncidentStyles);
 
   const alerts = incident.alerts;
   if (!alerts) {
@@ -767,7 +774,7 @@ function GroupedIncident({ incident, datetimeReference }: { incident: GroupedAle
   const [incidentRawResponse, setIncidentRawResponse] = useState<{ id: string; raw_request_data: any }>(undefined);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const payloadJSON = isModalOpen ? JSON.stringify(incidentRawResponse.raw_request_data, null, 4) : undefined;
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getIncidentStyles);
 
   return (
     <>
@@ -844,7 +851,7 @@ function AttachedIncidentsList({
   getUnattachClickHandler(pk: string): void;
 }) {
   const store = useStore();
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getIncidentStyles);
   const incident = store.alertGroupStore.alerts.get(id);
 
   if (!incident.dependent_alert_groups.length) {
@@ -880,7 +887,7 @@ function AttachedIncidentsList({
 }
 
 const AlertGroupStub = ({ buttons }: { buttons: React.ReactNode }) => {
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getIncidentStyles);
   return (
     <div className={styles.alertGroupStub}>
       <VerticalGroup align="center" spacing="md">
@@ -902,221 +909,6 @@ const AlertGroupStub = ({ buttons }: { buttons: React.ReactNode }) => {
   );
 };
 
-const getStyles = (theme: GrafanaTheme2) => {
-  return {
-    incidentRow: css`
-      display: flex;
-    `,
-
-    incidentRowLeftSide: css`
-      flex-grow: 1;
-    `,
-
-    block: css`
-      padding: 0 0 20px 0;
-    `,
-
-    payloadSubtitle: css`
-      margin-bottom: 16px;
-    `,
-
-    infoRow: css`
-      width: 100%;
-      border-bottom: 1px solid ${theme.colors.border.medium};
-      padding-bottom: 20px;
-    `,
-
-    buttonsRow: css`
-      margin-top: 20px;
-    `,
-
-    content: css`
-      margin-top: 5px;
-      display: flex;
-    `,
-
-    timelineIconBackground: css`
-      width: 28px;
-      height: 28px;
-      border-radius: 50%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      background: rgba(${theme.isDark ? '70, 76, 84, 1' : '70, 76, 84, 0'});
-    `,
-
-    message: css`
-      margin-top: 16px;
-      word-wrap: break-word;
-
-      a {
-        word-break: break-all;
-      }
-
-      ul {
-        margin-left: 24px;
-      }
-
-      p {
-        margin-bottom: 0;
-      }
-
-      code {
-        white-space: break-spaces;
-      }
-    `,
-
-    image: css`
-      margin-top: 16px;
-      max-width: 100%;
-    `,
-
-    collapse: css`
-      margin-top: 16px;
-      position: relative;
-    `,
-
-    column: css`
-      width: 50%;
-      padding-right: 24px;
-
-      &:not(:first-child) {
-        padding-left: 24px;
-      }
-    `,
-
-    incidentsContent: css`
-      > div:not(:last-child) {
-        border-bottom: 1px solid ${Colors.BORDER};
-        padding-bottom: 16px;
-      }
-
-      > div:not(:first-child) {
-        padding-top: 16px;
-      }
-    `,
-
-    timeline: css`
-      list-style-type: none;
-      margin: 0 0 24px 12px;
-    `,
-
-    timelineItem: css`
-      margin-top: 12px;
-    `,
-
-    notFound: css`
-      margin: 50px auto;
-      text-align: center;
-    `,
-
-    alertGroupStub: css`
-      margin: 24px auto;
-      width: 520px;
-      text-align: center;
-    `,
-
-    alertGroupStubDivider: css`
-      width: 520px;
-    `,
-
-    blue: css`
-      background: ${getLabelBackgroundTextColorObject('blue', theme).sourceColor};
-    `,
-
-    timelineTitle: css`
-      margin-bottom: 24px;
-    `,
-
-    timelineFilter: css`
-      margin-bottom: 24px;
-    `,
-
-    titleIcon: css`
-      color: ${theme.colors.secondary.text};
-      margin-left: 4px;
-    `,
-
-    integrationLogo: css`
-      margin-right: 8px;
-    `,
-
-    labelButton: css`
-      padding: 0 8px;
-      font-weight: 400;
-
-      &:disabled {
-        border: 1px solid ${theme.colors.border.strong};
-      }
-    `,
-
-    labelButtonText: css`
-      max-width: 160px;
-      overflow: hidden;
-      position: relative;
-      display: inline-block;
-      text-align: center;
-      text-decoration: none;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    `,
-
-    sourceName: css`
-      display: flex;
-      align-items: center;
-    `,
-
-    statusTagContainer: css`
-      margin-right: 8px;
-      display: inherit;
-    `,
-
-    statusTag: css`
-      height: 24px;
-      padding: 5px 8px;
-      border-radius: 2px;
-    `,
-
-    pagedUsers: css`
-      width: 100%;
-    `,
-
-    // TODO: Where are trash-button/hover-button coming from?
-    pagedUsersList: css`
-      list-style-type: none;
-      margin-bottom: 20px;
-      width: 100%;
-
-      & > li .trash-button {
-        display: none;
-      }
-
-      & > li:hover .trash-button {
-        display: block;
-      }
-
-      & > li {
-        padding: 8px 12px;
-        width: 100%;
-
-        & .hover-button {
-          display: none;
-        }
-      }
-
-      & > li:hover {
-        background: ${theme.colors.background.secondary};
-
-        & .hover-button {
-          display: inline-flex;
-        }
-      }
-    `,
-
-    userBadge: css`
-      vertical-align: middle;
-    `,
-  };
-};
-
-export const IncidentPage = withRouter(withMobXProviderContext(withTheme2(_IncidentPage)));
+export const IncidentPage = withRouter<RouteProps, IncidentPageProps>(
+  withMobXProviderContext(withTheme2(_IncidentPage))
+);
