+++ grafana-plugin/src/utils/hoc.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import { useDrawer } from './hooks';
+import { NavigateFunction, Params, useLocation, useNavigate, useParams } from 'react-router-dom';
 
 export const withDrawer = <T extends string>(Component: React.ComponentType<any>) => {
   const ComponentWithDrawer = (props: any) => {
@@ -9,3 +10,25 @@ export const withDrawer = <T extends string>(Component: React.ComponentType<any>
   };
   return ComponentWithDrawer;
 };
+
+interface Router<T> {
+  location: Location;
+  navigate: NavigateFunction;
+  params: Readonly<T>;
+}
+
+export interface PropsWithRouter<T> {
+  router: Router<T>;
+}
+
+export function withRouter<X, T extends PropsWithRouter<X>>(Component: React.FC<T>): React.FC<Omit<T, 'router'>> {
+  function HOCWithRouter(props: T) {
+    const location = useLocation();
+    const navigate = useNavigate();
+    const params = useParams() as unknown as X;
+
+    return <Component {...props} router={{ location, navigate, params }} />;
+  }
+
+  return HOCWithRouter;
+}
